+++ src/test/java/app/bottlenote/alcohols/integration/AlcoholQueryControllerIntegrationTest.java
@@ -1,30 +1,31 @@
 package app.bottlenote.alcohols.integration;
 
+import app.bottlenote.IntegrationTestSupport;
+import app.bottlenote.alcohols.domain.Alcohol;
+import app.bottlenote.alcohols.dto.response.detail.AlcoholDetail;
+import app.bottlenote.alcohols.fixture.AlcoholTestFactory;
+import app.bottlenote.global.data.response.GlobalResponse;
+import app.bottlenote.rating.fixture.RatingTestFactory;
+import app.bottlenote.user.domain.User;
+import app.bottlenote.user.domain.constant.SocialType;
+import app.bottlenote.user.dto.request.OauthRequest;
+import com.fasterxml.jackson.databind.JsonNode;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Tag;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.MediaType;
+import org.springframework.test.web.servlet.MvcResult;
+
+import java.nio.charset.StandardCharsets;
+import java.util.List;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.csrf;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
-
-import app.bottlenote.IntegrationTestSupport;
-import app.bottlenote.alcohols.domain.Alcohol;
-import app.bottlenote.alcohols.dto.response.detail.AlcoholDetail;
-import app.bottlenote.alcohols.fixture.AlcoholTestFactory;
-import app.bottlenote.global.data.response.GlobalResponse;
-import app.bottlenote.user.domain.User;
-import app.bottlenote.user.domain.constant.SocialType;
-import app.bottlenote.user.dto.request.OauthRequest;
-import com.fasterxml.jackson.databind.JsonNode;
-import java.nio.charset.StandardCharsets;
-import java.util.List;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Tag;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.MediaType;
-import org.springframework.test.context.jdbc.Sql;
-import org.springframework.test.web.servlet.MvcResult;
 
 
 @Tag("integration")
@@ -33,11 +34,10 @@ class AlcoholQueryControllerIntegrationTest extends IntegrationTestSupport {
 
 	@Autowired
 	AlcoholTestFactory alcoholTestFactory;
+	@Autowired
+	RatingTestFactory ratingTestFactory;
 
 	@DisplayName("알코올 상세 조회 시 해당 알코올을 마셔본 팔로잉 유저의 정보를 조회할 수 있다.")
-	@Sql(scripts = {
-		"/init-script/init-alcohol.sql",
-	})
 	@Test
 	void test_1() throws Exception {
 		// given
@@ -45,10 +45,11 @@ class AlcoholQueryControllerIntegrationTest extends IntegrationTestSupport {
 		final String userEmail = "test1@example.com";
 		final List<SocialType> userSocialType = List.of(SocialType.KAKAO);
 
+		Alcohol alcohol = alcoholTestFactory.createAlcohol();
 		User user1 = alcoholTestFactory.createUser(1L, userEmail, "test1");
 		User user2 = alcoholTestFactory.createUser(2L, "test@example.com", "test2");
-		alcoholTestFactory.createRating(user2, Alcohol.builder().id(1L).build(), 3);
 		alcoholTestFactory.createFollow(user1, user2);
+		ratingTestFactory.createRating(user2, alcohol, 3);
 
 		MvcResult authResult = mockMvc.perform(post("/api/v1/oauth/login")
 				.contentType(MediaType.APPLICATION_JSON)
@@ -78,4 +79,11 @@ class AlcoholQueryControllerIntegrationTest extends IntegrationTestSupport {
 		assertEquals(1, alcoholDetail.friendsInfo().getFollowerCount());
 		assertEquals(user2.getId(), alcoholDetail.friendsInfo().getFriends().get(0).userId());
 	}
+
+	/**
+	 * TODO : 추가해야 할 테스트 목록
+	 * 1. 노출 예상 사용자지만 사용자가 블럭됬을 경우
+	 * 2 노출 예상 사용자지만 사용자가 탈퇴한 경우(논리적 삭제 상태)
+	 * 3. 외 기본적인 오류 상황
+	 */
 }
